no task: fix update period serialization

+++ contracts/neutron_interchain_queries/src/contract.rs
@@ -206,7 +206,7 @@ pub fn register_transfers_query(
     connection_id: String,
     recipient: String,
     update_period: u64,
-    min_height: Option<u128>,
+    min_height: Option<u64>,
 ) -> NeutronResult<Response<NeutronMsg>> {
     let msg =
         new_register_transfers_query_msg(connection_id, recipient, update_period, min_height)?;
